refactor: extract 'HandlerUtils' to :common:android
Pure Android Handler/Looper helpers with zero project dependencies so they belong in :common:android, not :AnkiDroid. Helps untangle feature modules
that need main-thread posting utilities without pulling in :AnkiDroid.
